+++ packages/effect/src/internal/config.ts
@@ -582,8 +582,8 @@ export const validate = dual<
 
 /** @internal */
 export const withDefault = dual<
-  <A2>(def: A2) => <A>(self: Config.Config<A>) => Config.Config<A | A2>,
-  <A, A2>(self: Config.Config<A>, def: A2) => Config.Config<A | A2>
+  <const A2>(def: A2) => <A>(self: Config.Config<A>) => Config.Config<A | A2>,
+  <A, const A2>(self: Config.Config<A>, def: A2) => Config.Config<A | A2>
 >(2, (self, def) =>
   orElseIf(self, {
     orElse: () => succeed(def),