+++ packages/platform-node-shared/src/internal/fileSystem.ts
@@ -147,7 +147,7 @@ const removeFactory = (method: string) => {
   return (path: string, options?: FileSystem.RemoveOptions) =>
     nodeRm(
       path,
-      { recursive: options?.recursive ?? false }
+      { recursive: options?.recursive ?? false, force: options?.force ?? false }
     )
 }
 const remove = removeFactory("remove")
@@ -391,16 +391,11 @@ const makeTempFileScoped = (() => {
 
 // == readDirectory
 
-const readDirectory = (() => {
-  const nodeReadDirectory = effectify(
-    NFS.readdir,
-    handleErrnoException("FileSystem", "readDirectory"),
-    handleBadArgument("readDirectory")
-  )
-
-  return (path: string, options?: FileSystem.ReadDirectoryOptions) =>
-    nodeReadDirectory(path, options) as Effect.Effect<ReadonlyArray<string>, Error.PlatformError>
-})()
+const readDirectory = (path: string, options?: FileSystem.ReadDirectoryOptions) =>
+  Effect.tryPromise({
+    try: () => NFS.promises.readdir(path, options),
+    catch: (err) => handleErrnoException("FileSystem", "readDirectory")(err as any, [path])
+  })
 
 // == readFile